Add checking if the vaiable exists or not in all blade files, Add the writer info and the latest modified date in article view, Link the tags button to the article search
In various blade files, before looping and displaying the variables
returned from the controller, if the variable is null or not is checked.

The tags shown in article view are now also linked for the user to
easily search other aritcles having the same tag.

The writer information and the updated on date are shown in article
view.